% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addASE.R
\name{addASE}
\alias{addASE}
\title{Read ASE info and merge with vcf info. Check for mismatching alleles and prepare dataframe for manual phasing.}
\usage{
addASE(
  df_var,
  ASE_path,
  BM,
  reads = 2,
  od = 10,
  expand_region = 0,
  refAllele = "refAllele",
  altAllele = "altAllele",
  position = "position",
  refCount = "refCount",
  altCount = "altCount",
  contig = "contig",
  totalCount = "totalCount"
)
}
\arguments{
\item{df_var}{A df obtained from snps_read_and_merge()/conshap()}

\item{ASE_path}{A path of the ASE obtained from gex bulk analysis}

\item{BM}{A df for SNPs annotation}

\item{reads}{Number of minimum reads overlapping a SNPs}

\item{od}{overall depth, default ≥ 10}

\item{expand_region}{expand gene region by indicated bp}

\item{refAllele}{refAllele ASE}

\item{altAllele}{altAllele ASE}

\item{position}{position ASE}

\item{refCount}{refCount ASE}

\item{altCount}{altCount ASE}

\item{contig}{contig ASE}

\item{totalCount}{totalCount ASE}
}
\value{
A dataframe.
}
\description{
Read ASE info and merge with vcf info. Check for mismatching alleles and prepare dataframe for manual phasing.
}
