% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annhub_df.R
\name{annotationhub_df}
\alias{annotationhub_df}
\title{Extract annotation db from AnnotationHub}
\usage{
annotationhub_df(edb_id = "AH113665", window = 2.5)
}
\arguments{
\item{edb_id}{ensembl database id from AnnotationHub i.e. "AH113665"}

\item{window}{Window size up and downstream gene size.}
}
\value{
A dataframe.
}
\description{
Extract annotation db from AnnotationHub
}
