% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biall_def.R
\name{biallmonoall}
\alias{biallmonoall}
\title{Function for the definition of biallelic/monoallelic SNPs based on allelic counts}
\usage{
biallmonoall(
  df,
  refFracValue = c(0.15, 0.85),
  biallObs = 1,
  monoallObs = 1,
  bt = 2,
  br = 2
)
}
\arguments{
\item{df}{A dataframe with columns variantID, ensembl_gene_id, hgnc_symbol, refFrac.}

\item{refFracValue}{refFrac values where to cut biall/monoall default 0.15,0.85}

\item{biallObs}{Minimum number of SNPs supporting biallelic observation}

\item{monoallObs}{Minimum number of SNPs supporting monoallelic observation}

\item{bt}{biallelic threshold (min number of SNPs to consider a gene always biallelic)}

\item{br}{min number of reads on minor allele for biallelic definition}
}
\value{
A df
}
\description{
Function for the definition of biallelic/monoallelic SNPs based on allelic counts
}
