% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_df.R
\name{biomart_df}
\alias{biomart_df}
\title{Create annotation df from biomart}
\usage{
biomart_df(dataset_choice = "hsapiens_gene_ensembl", window = 2.5)
}
\arguments{
\item{dataset_choice}{A string with mart dataset.}

\item{window}{Window size up and downstream gene size.}
}
\value{
A dataframe.
}
\description{
Create annotation df from biomart
}
