% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeP.R
\name{conshap}
\alias{conshap}
\title{Read and hapotype phasing results from Shapeit, Haptreex,...}
\usage{
conshap(
  df1,
  df2,
  BMsel,
  CHROM = "CHROM",
  POS = "POS",
  Ref = "Ref",
  REF = "REF",
  ALT = "ALT",
  DF1 = "DF1",
  DF2 = "DF2"
)
}
\arguments{
\item{df1}{First dataframe with haplotype calling results}

\item{df2}{Second dataframe with haplotype calling results}

\item{BMsel}{BM dataframe containing only chromosome under analysis}

\item{CHROM}{Chromosome column}

\item{POS}{Position column}

\item{Ref}{Reference haplotype column}

\item{REF}{Reference allele column}

\item{ALT}{Alternative allele column}

\item{DF1}{Name of the tool used for the haplotype calling in df1}

\item{DF2}{Name of the tool used for the haplotype calling in df2}
}
\value{
A dataframe.
}
\description{
Read and hapotype phasing results from Shapeit, Haptreex,...
}
