% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_alleles.R
\name{flip_alleles}
\alias{flip_alleles}
\title{Flip given alleles}
\usage{
flip_alleles(df_toswap, col1, col2, rownum)
}
\arguments{
\item{df_toswap}{A dataframe}

\item{col1}{Col name first allele to swap}

\item{col2}{Col name second allele to swap}

\item{rownum}{Row num to swap}
}
\value{
A dataframe.
}
\description{
Flip given alleles
}
