% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_phasing.R
\name{manualP}
\alias{manualP}
\title{Apply manual phasing}
\usage{
manualP(df)
}
\arguments{
\item{df}{A df containg annotated SNPs to phase (and phased)}
}
\value{
A dataframe.
}
\description{
Apply manual phasing
}
