% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ase.R
\name{plot_annotation}
\alias{plot_annotation}
\title{Plot functional annotation of ASE/ASA analysis results. Plotting function from https://r-graph-gallery.com/297-circular-barplot-with-groups.html}
\usage{
plot_annotation(df, ASC, gene, annotation)
}
\arguments{
\item{df}{A dataframe.}

\item{ASC}{Allele Specific categorization in "biallelic/monoallelic/other" column identifier.}

\item{gene}{Gene-id column identifier.}

\item{annotation}{Functional annotation column to plot}
}
\value{
A plot
}
\description{
Plot functional annotation of ASE/ASA analysis results. Plotting function from https://r-graph-gallery.com/297-circular-barplot-with-groups.html
}
