% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ase.R
\name{plot_ase}
\alias{plot_ase}
\title{Read and merge vcf containing all and only (Shapeit4) phased variants}
\usage{
plot_ase(df, genes_to_plot, title)
}
\arguments{
\item{df}{A dataframe.}

\item{genes_to_plot}{A dataframe.}

\item{title}{Title for the plot}
}
\value{
A ggplot
}
\description{
Read and merge vcf containing all and only (Shapeit4) phased variants
}
