% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ase.R
\name{plot_msigdb}
\alias{plot_msigdb}
\title{Read and merge vcf containing all and only (Shapeit4) phased variants}
\usage{
plot_msigdb(
  df,
  msigdb_species = "human",
  msigdb_cat = "C5",
  msigdb_subcat = "BP",
  ASC = "type",
  gene = "ensembl_gene_id"
)
}
\arguments{
\item{df}{A dataframe.}

\item{msigdb_species}{species.}

\item{msigdb_cat}{Category for msibdg, see msigdbr_collections()}

\item{msigdb_subcat}{msigdb subcategory}

\item{ASC}{Allele Specific categorization in "biallelic/monoallelic/other" column identifier in df.}

\item{gene}{Gene-id column identifier in df.}
}
\value{
A ggplot
}
\description{
Read and merge vcf containing all and only (Shapeit4) phased variants
}
