% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_prep.R
\name{prep_outcome}
\alias{prep_outcome}
\title{Read and merge vcf containing all and only (Shapeit4) phased variants}
\usage{
prep_outcome(df_initial, df_phased)
}
\arguments{
\item{df_initial}{Dataframe originated by conshap() or snps_read_merge()/_haptreex()}

\item{df_phased}{Result of monoallelic() function}
}
\value{
A dataframe.
}
\description{
Read and merge vcf containing all and only (Shapeit4) phased variants
}
