% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon.R
\name{read_salmon}
\alias{read_salmon}
\title{Read and merge vcf containing all and only (Shapeit4) phased variants}
\usage{
read_salmon(salmon_path, txdb, counts)
}
\arguments{
\item{salmon_path}{A character vector with quant.sf path.}

\item{txdb}{txdb for salmon import}

\item{counts}{Filtering value on salmon counts}
}
\value{
A dataframe.
}
\description{
Read and merge vcf containing all and only (Shapeit4) phased variants
}
