% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps_read_merge.R
\name{read_shapeit}
\alias{read_shapeit}
\title{Read and merge vcf containing all and only (Shapeit4) phased variants}
\usage{
read_shapeit(
  variant_calling_out_path,
  phased_out_path,
  sample_name = "SAMPLE1"
)
}
\arguments{
\item{variant_calling_out_path}{A character vector with vcf path.}

\item{phased_out_path}{A character vector with phased vcf path.}

\item{sample_name}{A character vector with sample identifier in vcf.}
}
\value{
A dataframe.
}
\description{
Read and merge vcf containing all and only (Shapeit4) phased variants
}
