% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon.R
\name{seesaw}
\alias{seesaw}
\title{SEESAW analysis plotting}
\usage{
seesaw(names, edb_id = "AH113665", tss_gap = 50, gene_id, files, cellline_id)
}
\arguments{
\item{names}{A character vector with the sample id}

\item{edb_id}{ensembl database id from AnnotationHub i.e. "AH113665"}

\item{tss_gap}{gap size for tss grouping}

\item{gene_id}{ensebl gene id to plot}

\item{files}{path to quant.sf file from the seesaw pipeline}

\item{cellline_id}{cellular line description i.e. "naive"}
}
\value{
A plot.
}
\description{
SEESAW analysis plotting
}
