% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_df.R
\name{biomart_df}
\alias{biomart_df}
\title{Read and merge vcf containing all and only (Shapeit4) phased variants}
\usage{
biomart_df(dataset_choice = "hsapiens_gene_ensembl", window = 2.5)
}
\arguments{
\item{dataset_choice}{A string with mart dataset.}

\item{window}{Window size up and downstream gene size.}
}
\value{
A dataframe.
}
\description{
Read and merge vcf containing all and only (Shapeit4) phased variants
}
\examples{
biomart_df("hsapiens_gene_ensembl")
}
